Add a PHP 8 attribute to support configuration
